+++ irelia/examples/rest.rs
@@ -4,7 +4,7 @@ extern crate tokio;
 #[tokio::main]
 async fn main() {
     let request_client = irelia::RequestClient::new();
-    let lcu_client = irelia::rest::LCUClient::new().unwrap();
+    let lcu_client = irelia::rest::LcuClient::new(false).unwrap();
 
     let value: serde_json::Value = lcu_client
         .get("Example", &request_client)